requirements.txt
# Dev environment
pip
autopep8

# App
flask
Flask-SQLAlchemy
python-dotenv
